It's works pretty nice :) TODO: beter loging

+++ src/daemon/service_routines.c
@@ -6,7 +6,7 @@
 /*   By: ahrytsen <ahrytsen@student.unit.ua>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/07/14 22:35:53 by ahrytsen          #+#    #+#             */
-/*   Updated: 2018/07/18 17:33:46 by ahrytsen         ###   ########.fr       */
+/*   Updated: 2018/07/18 21:09:25 by ahrytsen         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,20 +21,16 @@ static int	proc_watch(t_job *job)
 		pthread_mutex_lock(&job->jmutex);
 		time(&job->t);
 		job->ex_st = st;
-		if (WIFSTOPPED(st))
+		if (WIFSTOPPED(st) && job->proc->stopsignal == WSTOPSIG(st))
 		{
-			if (job->proc->stopsignal == WSTOPSIG(st))
-			{
-				job->status = stoping;
-				sleep(job->proc->stoptime);
-				kill(job->pid, SIGKILL);
-			}
+			job->status = stoping;
+			sleep(job->proc->stoptime);
+			kill(job->pid, SIGKILL);
 		}
-		else
+		else if (!WIFSTOPPED(st))
 		{
 			st = WEXITSTATUS(st);
 			job->status = stop;
-			ft_dprintf(2, "exit: %d\n", st);
 			return ((job->proc->autorestart == always
 					|| (job->proc->autorestart == unexp
 						&& st < 256 && !job->proc->exitcodes[st])) ? 0 : 1);
@@ -66,21 +62,22 @@ static int	create_chld(t_job *job)
 		job->proc->stderr ? 0 : close(job->p_err[1]);
 		proc_start_chld(job);
 	}
-	job->proc->stdin ? 0 : close(job->p_in[0]);
-	job->proc->stdout ? 0 : close(job->p_out[0]);
-	job->proc->stderr ? 0 : close(job->p_err[0]);
 	pthread_mutex_unlock(&job->jmutex);
 	pthread_mutex_unlock(&job->proc->pmutex);
 	return (ret);
 }
 
-static int proc_spawn(t_job *job)
+static int	proc_spawn(t_job *job)
 {
 	int	st;
 
 	job->start_tries = 0;
 	while (job->start_tries <= job->proc->startretries)
 	{
+		pthread_mutex_lock(&get_dconf()->dmutex);
+		ft_dprintf(2, "%s:%zu starting...(attempt %d/%d)\n", job->proc->name,
+			job - job->proc->jobs, job->start_tries, job->proc->startretries);
+		pthread_mutex_unlock(&get_dconf()->dmutex);
 		st = create_chld(job);
 		pthread_mutex_lock(&get_dconf()->dmutex);
 		pthread_mutex_lock(&job->jmutex);
@@ -98,7 +95,7 @@ static int proc_spawn(t_job *job)
 	return (-1);
 }
 
-void	*proc_service(void *data)
+void		*proc_service(void *data)
 {
 	t_job	*job;
